FIX: Solved RTCP Padding problem
Previous value of FIELD_MIN_LENGTH was 2, which was certainly incorrect.
Each TLV field has to be multiple of 4, thus minimal size is 4, and ID
(8 bits), field length (8 bits) and a padding of another 8 bits to pad the field to the nearest multiple of 4 bytes. Resulting effectively in 4 bytes.

rtcp_padding(...) was also added in order to check for padding in a reusable manner. This function can also be exported to a common file, in order to be shared between several related dissectors (MCvideo, MCdata...)