FIX: common.lua file was not found in some platforms
It seems that the plugin execution folder is not the same as the folder where personal plugins are stored in some platforms. In both mcptt.lua and mcvideo.lua scripts doFile(...) function crashed because mcx-common/common.lua script could not be found. So instead of using relative paths, absolute paths will be used.
The directory separator has also been defined as platform specific.