REFACTOR: Fixed and improved inclusion/installation path for all users
Due to different wireshark plugins folders and ease of installation for other users, it was decided the installation path to be `~/.config/wireshark/plugins`.

A install script was written to assist the installation or upgrade of the MCX Dissector and the relevant files were refactored to include the right files.